refactor: split textbook code block rows
